feat(robberOps): fix mathod and make their return type optionals
