+++ src/main/scala/scatan/model/game/ops/RobberOps.scala
@@ -2,6 +2,7 @@ package scatan.model.game.ops
 
 import scatan.model.game.ScatanState
 import scatan.model.map.Hexagon
+import scatan.model.map.HexagonInMap.layer
 
 object RobberOps:
   extension (state: ScatanState)
@@ -10,6 +11,8 @@ object RobberOps:
       * @param hexagon
       *   the hexagon to move the robber to
       * @return
-      *   a new ScatanState with the robber moved to the specified hexagon
+      *   Some(ScatanState) if the robber was moved, None otherwise
       */
-    def moveRobber(hexagon: Hexagon): ScatanState = state.copy(robberPlacement = hexagon)
+    def moveRobber(hexagon: Hexagon): Option[ScatanState] =
+      if state.robberPlacement == hexagon || hexagon.layer > state.gameMap.withTerrainLayers then None
+      else Some(state.copy(robberPlacement = hexagon))
